FADEL 7war 2l phrase

+++ src/com/APT/SearchEngine/PageRanking/PageRanking.java
@@ -10,13 +10,13 @@ import java.util.stream.Stream;
 
 public class PageRanking {
 
-    private ArrayList<HashMap<String, Integer>> searchedQuery = new ArrayList<HashMap<String, Integer>>();
+    private ArrayList<HashMap<String, Float>> searchedQuery = new ArrayList<HashMap<String, Float>>();
 
-    public ArrayList<HashMap<String, Integer>> getSearchedQuery() {
+    public ArrayList<HashMap<String, Float>> getSearchedQuery() {
         return searchedQuery;
     }
 
-    public void setSearchedQuery(ArrayList<HashMap<String, Integer>> searchedQuery) {
+    public void setSearchedQuery(ArrayList<HashMap<String, Float>> searchedQuery) {
         this.searchedQuery = searchedQuery;
     }
 
@@ -24,19 +24,19 @@ public class PageRanking {
     public PageRanking() {
     }
 
-    public ArrayList<String> RankerNormal(ArrayList<HashMap<String, Integer>> searchedQuery) {
+    public ArrayList<String> RankerNormal(ArrayList<HashMap<String, Float>> searchedQuery) {
 
         setSearchedQuery(searchedQuery);
-        Map<String, Integer> Union = new HashMap<String, Integer>();
+        Map<String, Float> Union = new HashMap<String, Float>();
 
         Union.putAll(searchedQuery.get(0));
 
 
         for (int i = 1; i < searchedQuery.size(); i++) {
 
-            for (Map.Entry<String, Integer> entry : searchedQuery.get(i).entrySet()) {
+            for (Map.Entry<String, Float> entry : searchedQuery.get(i).entrySet()) {
                 String key = entry.getKey();
-                int value = entry.getValue();
+                Float value = entry.getValue();
 
                 if (!Union.containsKey(key)) {
                     Union.put(key, value);
@@ -47,10 +47,10 @@ public class PageRanking {
 
             }
         }
-        Stream<Map.Entry<String, Integer>> sorted = Union.entrySet().stream().sorted(Map.Entry.comparingByValue());
+        Stream<Map.Entry<String, Float>> sorted = Union.entrySet().stream().sorted(Map.Entry.comparingByValue());
 
 
-        Map<String, Integer> tempMap =
+        Map<String, Float> tempMap =
                 Union.entrySet().stream()
                         .sorted(Map.Entry.comparingByValue(Comparator.reverseOrder()))
                         .collect(Collectors.toMap(
@@ -65,31 +65,31 @@ public class PageRanking {
     /*
      ArrayList of HashMaps, each HashMaps represents a word
      */
-    public ArrayList<String> PhraseRanker(ArrayList<HashMap<String, Pair<Integer, ArrayList<String>>>> searchedQuery, ArrayList<Pair<String, Integer>> IndexOfPhrase) {
+    public ArrayList<String> PhraseRanker(ArrayList<HashMap<String, Pair<Float, ArrayList<String>>>> searchedQuery, ArrayList<Pair<String, Integer>> IndexOfPhrase) {
 
-        HashMap<String, Integer> RankedList = new HashMap<>();
-        HashMap<String, ArrayList<Integer>> ExpectedLocationOfWords = new HashMap<>();
+        HashMap<String, Float> RankedList = new HashMap<>();
+        HashMap<String, ArrayList<Float>> ExpectedLocationOfWords = new HashMap<>();
         boolean firstWord = true;
         ArrayList<Integer> TempRanker = new ArrayList<>();
         //Considering each and every link alone
 
-        HashMap<String, Pair<Integer, ArrayList<String>>> Object = searchedQuery.get(0);
+        HashMap<String, Pair<Float, ArrayList<String>>> Object = searchedQuery.get(0);
         boolean found  = false;
         //Looping on the URLS of the first word to add the ArrayList of ints instead of words
-        for (Map.Entry<String,Pair<Integer, ArrayList<String>>> item : Object.entrySet())
+        for (Map.Entry<String,Pair<Float, ArrayList<String>>> item : Object.entrySet())
         {
             RankedList.put(item.getKey(),item.getValue().getKey());
-            ArrayList<Integer> Temp= new ArrayList<>();
+            ArrayList<Float> Temp= new ArrayList<>();
             for(String Number: item.getValue().getValue())
             {
-                Temp.add(Integer.parseInt(Number));
+                Temp.add(Float.parseFloat(Number));
             }
             ExpectedLocationOfWords.put(item.getKey(),new ArrayList<>(Temp));
         }
 
         int WordNum=0;
         //Looping on all the words now to make sure they match the criteria
-        for(HashMap<String, Pair<Integer, ArrayList<String>>> item: searchedQuery)
+        for(HashMap<String, Pair<Float, ArrayList<String>>> item: searchedQuery)
         //Holding the hashmap that refers to each word alone
         {
             if(firstWord)
@@ -100,7 +100,7 @@ public class PageRanking {
             }
 
             //Loop on all the occurrences of this word in this URL
-            for(Map.Entry<String, Pair<Integer, ArrayList<String>>> HashMapsURL: item.entrySet())
+            for(Map.Entry<String, Pair<Float, ArrayList<String>>> HashMapsURL: item.entrySet())
             {
                 //Inside the URL i have both Rank and string of occurrences, loop on the
                 //indexes and if it happens that this word is found in the correct position then add the rank to the pre-saved rank
@@ -147,7 +147,7 @@ public class PageRanking {
 
 
 
-        Map<String, Integer> tempMap =
+        Map<String, Float> tempMap =
                 RankedList.entrySet().stream()
                         .sorted(Map.Entry.comparingByValue(Comparator.reverseOrder()))
                         .collect(Collectors.toMap(
